Use bigint for Telegram user ids

+++ alembic/versions/20260511_0001_initial_schema.py
@@ -24,7 +24,7 @@ def upgrade() -> None:
         op.create_table(
             "users",
             sa.Column("id", sa.Integer(), nullable=False),
-            sa.Column("telegram_id", sa.Integer(), nullable=False),
+            sa.Column("telegram_id", sa.BigInteger(), nullable=False),
             sa.Column("telegram_username", sa.String(length=255), nullable=True),
             sa.Column("first_name", sa.String(length=255), nullable=True),
             sa.Column("last_name", sa.String(length=255), nullable=True),
